quick fixes/changes

+++ app/src/main/java/com/example/icecaremobile/domain/model/Response/LoginResponse.kt
@@ -16,7 +16,7 @@ data class Response(
     val status: String,
     val accountBalance: String,
     val accountNumber: String,
-    val dollarRate: Long,
+    val dollarRate: Double,
     val companyNumber: String,
     val companyAccounts: List<CompanyAccounts>
 )
